+++ src/transactions/SignatureUtils.cpp
@@ -40,7 +40,7 @@ signHashX(const ByteSlice& x)
 {
     DecoratedSignature result;
     Signature out(0, 0);
-    out.resize(x.size());
+    out.resize(static_cast<uint32_t>(x.size()));
     std::memcpy(out.data(), x.data(), x.size());
     result.signature = out;
     result.hint = getHint(sha256(x));
